Move active item tests to a separate file
